backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=1.10.7
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6
asyncpg>=0.27.0
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Redis
redis>=4.5.4

# WebSockets
websockets>=11.0.2

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0
pyyaml>=6.0.0
